refactor: 로또 수익금 관련 변수를 Int에서 Long으로 변경

+++ src/main/kotlin/lotto/view/OutputView.kt
@@ -20,7 +20,7 @@ class OutputView {
         }
     }
 
-    fun printRevenueToCostRatio(revenue: Int, cost: Int) {
+    fun printRevenueToCostRatio(revenue: Long, cost: Int) {
         println(MESSAGE_PROFIT_TO_COST_RATIO.format(revenue / cost.toDouble() * 100))
     }
